Move export features to export service
